Checks bei Mitglied von Control nach Impl verschoben (#1042)
* zwischenstand

* Zwischenstand

* Migrate Mitgliedstyp

* Eine Datei nicht gespeichert

* Update

* FIx ID

* Update MitgliedControl.java

* Check auch für MandatsID

* Update Mitglied Import damit keine Änderung bei Verlassen angezeigt wird

* Mitgliedstyp geändert

---------

+++ src/de/jost_net/JVerein/io/Migration.java
@@ -590,7 +590,7 @@ public class Migration
       final Map<String, Integer> beitragsGruppen)
       throws RemoteException, SQLException, ApplicationException, ParseException
   {
-    m.setMitgliedstyp(Mitgliedstyp.MITGLIED);
+    m.setMitgliedstyp(Long.valueOf(Mitgliedstyp.MITGLIED));
 
     /*
      * necessary columns
@@ -899,19 +899,19 @@ public class Migration
     {
       if (mitgliedstyp.matches("[0-9]+"))
       {
-        m.setMitgliedstyp(Integer.parseInt(mitgliedstyp));
+        m.setMitgliedstyp(Long.parseLong(mitgliedstyp));
       }
       else
       {
         progMonitor.log(String.format(
             "Mitgliedstyp bei: %s ist entweder leer oder besteht nicht nur aus Zahlen, setze auf 1 (Mitglied)",
             Adressaufbereitung.getNameVorname(m)));
-        m.setMitgliedstyp(Mitgliedstyp.MITGLIED);
+        m.setMitgliedstyp(Long.valueOf(Mitgliedstyp.MITGLIED));
       }
     }
     else
     { // Default value
-      m.setMitgliedstyp(Mitgliedstyp.MITGLIED);
+      m.setMitgliedstyp(Long.valueOf(Mitgliedstyp.MITGLIED));
     }
 
     m.setVermerk1(getResultFrom(results, InternalColumns.VERMERKA));
